requirements.txt
discord.py==1.6.0
python-digitalocean==1.16.0

